Changed int to uint32_t

+++ kernel/arch/i386/source/kstdio.c
@@ -4,7 +4,7 @@ void printk(const char *__str)
 {
     uint16_t *vidmem = (uint16_t *)0xb8000;
 
-    for (int i = 0; __str[i] != '\0'; i++)
+    for (uint16_t i = 0; __str[i] != '\0'; i++)
     {
         vidmem[i] = (vidmem[i] & 0xFF00) | __str[i];
     }
